+++ Socolin.RabbitMQ.Client/RabbitMqEnqueueQueueClient.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Socolin.RabbitMQ.Client.Pipes.Client;
 using Socolin.RabbitMQ.Client.Pipes.Client.Context;
@@ -7,7 +8,7 @@ namespace Socolin.RabbitMQ.Client
 {
 	public interface IRabbitMqEnqueueQueueClient
 	{
-		Task EnqueueMessageAsync(object message);
+		Task EnqueueMessageAsync(object message, Dictionary<string, object>? contextItems = null);
 	}
 
 	public class RabbitMqEnqueueQueueClient : IRabbitMqEnqueueQueueClient
@@ -19,9 +20,9 @@ namespace Socolin.RabbitMQ.Client
 			_pipeline = pipeline;
 		}
 
-		public async Task EnqueueMessageAsync(object message)
+		public async Task EnqueueMessageAsync(object message, Dictionary<string, object>? contextItems = null)
 		{
-			var pipeMessage = new ClientPipeContextMessage(message);
+			var pipeMessage = new ClientPipeContextMessage(message, contextItems);
 			await ClientPipe.ExecutePipelineAsync(pipeMessage, _pipeline);
 		}
 	}